utils: rename pipeline.py -> io.py
The module holds file-list discovery, path builders, and the mini-database
(I/O helpers), so io.py is a clearer name than pipeline.py. Update all
importers (modules, recipes, tests); no behavior change.
